pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pdf-malware-scanner"
version = "0.1.0"
description = "Defensive PDF malware triage scanner with reports, evidence storage, and optional YARA/ClamAV integrations."
readme = "DESIGN_PLAN.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "PDF Scanner Maintainers" }]
dependencies = []

[project.optional-dependencies]
deep = ["pikepdf>=8.0"]
yara = ["yara-python>=4.5"]
test = ["pytest>=8.0"]

[project.scripts]
pdfscan = "pdfscan.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["pdfscan"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
